Revise FlatBuffers codec: consolidate 7 files into 3, eliminate wrapper types
Reduce the codec module from 2034 lines across 7 files to 1776 lines across 3
files. Eliminate 4 DecodedXxxTask wrapper structs in favor of a DecodedTask enum
holding fb types directly, merge DecodedJobInfo/DecodedAttempt into JobView/
JobAttemptView, and return owned types for small scalar records (JobStatus,
HolderRecord, JobCancellation). Update all consumer files for the new API.
