+++ src/job_attempt.rs
@@ -1,7 +1,5 @@
-use bytes::Bytes;
-
-use crate::codec::{DecodedAttempt, decode_attempt_bytes};
-use crate::job_store_shard::JobStoreShardError;
+// Re-export the codec's zero-copy JobAttemptView directly.
+pub use crate::codec::JobAttemptView;
 
 /// Outcome passed by callers when reporting an attempt's completion.
 #[derive(Debug, Clone)]
@@ -50,52 +48,4 @@ pub struct JobAttempt {
     /// or preserved from the source system for imported attempts.
     pub started_at_ms: i64,
     pub status: AttemptStatus,
-}
-
-/// Zero-copy view over serialized `JobAttempt` data.
-#[derive(Clone)]
-pub struct JobAttemptView {
-    decoded: DecodedAttempt,
-}
-
-impl std::fmt::Debug for JobAttemptView {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("JobAttemptView")
-            .field("job_id", &self.job_id())
-            .field("attempt_number", &self.attempt_number())
-            .field("task_id", &self.task_id())
-            .finish()
-    }
-}
-
-impl JobAttemptView {
-    pub fn new(bytes: impl Into<Bytes>) -> Result<Self, JobStoreShardError> {
-        // Validate and decode up front; reject invalid data early.
-        let decoded = decode_attempt_bytes(bytes.into())?;
-        Ok(Self { decoded })
-    }
-
-    pub fn job_id(&self) -> &str {
-        self.decoded.job_id()
-    }
-
-    pub fn attempt_number(&self) -> u32 {
-        self.decoded.attempt_number()
-    }
-
-    pub fn relative_attempt_number(&self) -> u32 {
-        self.decoded.relative_attempt_number()
-    }
-
-    pub fn task_id(&self) -> &str {
-        self.decoded.task_id()
-    }
-
-    pub fn started_at_ms(&self) -> i64 {
-        self.decoded.started_at_ms()
-    }
-
-    pub fn state(&self) -> AttemptStatus {
-        self.decoded.state()
-    }
 }
